move bank account set payment to parent, use processing payment for polling
