+++ src/Apps/Order/Components/PollAccountBalance.tsx
@@ -1,7 +1,5 @@
 import { FC, useState } from "react"
 import { graphql, createRefetchContainer, RelayRefetchProp } from "react-relay"
-import { Spinner } from "@artsy/palette"
-import styled from "styled-components"
 
 import { useSystemContext } from "System"
 import { SystemQueryRenderer } from "System/Relay/SystemQueryRenderer"
@@ -12,6 +10,7 @@ import {
   PollAccountBalanceQueryResponse,
 } from "__generated__/PollAccountBalanceQuery.graphql"
 import { PollAccountBalance_commerceBankAccountBalance } from "__generated__/PollAccountBalance_commerceBankAccountBalance.graphql"
+import { ProcessingPayment } from "./ProcessingPayment"
 
 interface PollAccountBalanceProps {
   relay: RelayRefetchProp
@@ -71,7 +70,7 @@ const PollAccountBalance: FC<PollAccountBalanceProps> = ({
     clearTimeout(timeoutID)
   }
 
-  return <PollAccountBalancePlaceholder />
+  return <ProcessingPayment />
 }
 
 export const BALANCE_QUERY = graphql`
@@ -112,7 +111,7 @@ export const PollAccountBalanceQueryRenderer: FC<PollAccountBalanceQueryRenderer
   return (
     <SystemQueryRenderer<PollAccountBalanceQuery>
       environment={relayEnvironment}
-      placeholder={<PollAccountBalancePlaceholder />}
+      placeholder={<ProcessingPayment />}
       variables={{ setupIntentId }}
       query={BALANCE_QUERY}
       render={renderWithLoadProgress<PollAccountBalanceQueryResponse>(
@@ -125,18 +124,5 @@ export const PollAccountBalanceQueryRenderer: FC<PollAccountBalanceQueryRenderer
         )
       )}
     />
-  )
-}
-
-const PollAccountBalancePlaceholder: FC = () => {
-  const SpinnerContainer = styled.div`
-    width: 100%;
-    height: 300px;
-    position: relative;
-  `
-  return (
-    <SpinnerContainer data-testid="account-balance-placeholder">
-      <Spinner position="relative" color="brand" />
-    </SpinnerContainer>
   )
 }
